refactor: 불필요한 depth를 만드는 클래스 참조 삭제
