+++ src/main/java/racing/RacingGameManager.java
@@ -15,6 +15,7 @@ public class RacingGameManager {
 	private static final String REGEX = ",";
 	private static final int MAX_SIZE_FOR_CAR_NAME = 5;
 	private static final int STANDARD_RANGE_NUMBER = 10;
+	private final static int MIN_CONDITION_FOR_MOVING = 4;
 
 	private CarFactory carFactory;
 
@@ -22,7 +23,7 @@ public class RacingGameManager {
 		this.carFactory = new CarFactory();
 	}
 
-	public void createCarList(String inputData) {
+	void createCarList(String inputData) {
 		List<String> splitList = CustomStringUtils.split(inputData, REGEX);
 		for(String value : splitList){
 			carFactory.add(makeCar(value));
@@ -36,16 +37,23 @@ public class RacingGameManager {
 		return new Car(value);
 	}
 
-	public void playGames(final GameCommand gameCommand){
+	void playGames(final GameCommand gameCommand){
 		while(gameCommand.isPlaying()){
 			movingCars();
 			ConsoleUtils.printLine();
 		}
 	}
 
-	public void movingCars(){
+	private void movingCars(){
 		for(Car car : carFactory.getCarList()){
-			RacingGame.play(car, RandonNumberUtils.generate(STANDARD_RANGE_NUMBER));
+			play(car, RandonNumberUtils.generate(STANDARD_RANGE_NUMBER));
 		}
+	}
+
+	void play(Car car, int i) {
+		if(i >= MIN_CONDITION_FOR_MOVING){
+			car.go();
+		}
+		ConsoleUtils.print(car.makeResult());
 	}
 }
